refactor some types
