+++ packages/remirror__extension-positioner/src/positioner-extension.ts
@@ -19,16 +19,7 @@ import type { CreateEventHandlers } from '@remirror/extension-events';
 import { Decoration, DecorationSet } from '@remirror/pm/view';
 import { ExtensionPositionerTheme } from '@remirror/theme';
 
-import {
-  alwaysPositioner,
-  blockNodePositioner,
-  cursorPositioner,
-  emptyBlockNodeEndPositioner,
-  emptyBlockNodePositioner,
-  emptyBlockNodeStartPositioner,
-  nearestWordPositioner,
-  selectionPositioner,
-} from './core-positioners';
+import { positioners } from './core-positioners';
 import type {
   BasePositionerProps,
   Positioner,
@@ -254,48 +245,6 @@ export interface PositionerHandler {
  * positioner. Once teh
  */
 export type PositionerChangeHandlerMethod = (elementSetters: SetActiveElement[]) => void;
-
-export const positioners = {
-  /**
-   * Creates a rect which wraps the current selection.
-   */
-  selection: selectionPositioner,
-
-  /**
-   * Creates a rect for the cursor. Is inactive for
-   */
-  cursor: cursorPositioner,
-
-  /**
-   * Creates a positioner which always shows the position of the selection whether empty or not.
-   */
-  always: alwaysPositioner,
-
-  /**
-   * Creates a position which wraps the entire selected block node.
-   */
-  block: blockNodePositioner,
-
-  /**
-   * Creates a position which wraps the entire selected block node. This is only active when the block node is empty.
-   */
-  emptyBlock: emptyBlockNodePositioner,
-
-  /**
-   * Creates a position which wraps the entire selected block node. This is only active when the block node is empty.
-   */
-  emptyBlockStart: emptyBlockNodeStartPositioner,
-
-  /**
-   * Creates a position which wraps the entire selected block node. This is only active when the block node is empty.
-   */
-  emptyBlockEnd: emptyBlockNodeEndPositioner,
-
-  /**
-   * Create a rect which surrounds the nearest word.
-   */
-  nearestWord: nearestWordPositioner,
-};
 
 /**
  * This is a helper method for getting the positioner. The props can either
